Deletes unnecessary files, put a few classes in one file
