--- lib/vsge_internal_rectangle.hpp
@@ -1,15 +1,0 @@
-#ifndef VSGE_INTERNAL_RECTANGLE_HPP
-#define VSGE_INTERNAL_RECTANGLE_HPP
-
-#include "vsge_internal_shape.hpp"
-
-namespace vsge {
-	class Internal_Rectangle : public Internal_Shape {
-
-		public: //maybe "friend" in a future
-
-		Internal_Rectangle();
-	};
-}
-
-#endif
