--- lib/vsge_rectangle.cpp
@@ -1,9 +1,0 @@
-#include "vsge_rectangle.hpp"
-#include "vsge_core.hpp"
-
-namespace vsge {
-
-	Rectangle::Rectangle(int layer) {
-		internal = Core::getInstance().Rectangle_Factory(layer);
-	}
-}
