--- lib/vsge_rectangle.hpp
@@ -1,16 +1,0 @@
-#ifndef VSGE_RECTANGLE_HPP
-#define VSGE_RECTANGLE_HPP
-
-#include "vsge_shape.hpp"
-
-namespace vsge {
-
-	class Rectangle : public Shape {
-
-		public: 
-		Rectangle(int layer = 0);
-	};
-
-}
-
-#endif
