+++ lib/vsge_shapes.hpp
@@ -5,11 +5,9 @@
 
 namespace vsge
 {
-
 	class Internal_Shape;
 
-	class Shape
-	{
+	class Shape {
 
 	protected:
 		Internal_Shape *internal;
@@ -17,10 +15,15 @@ namespace vsge
 	public:
 		Shape();
 		virtual ~Shape();
-		void setPosition(Vector2f pos); // to replace with vsge point
+		void setPosition(Vector2f pos);
 		void setColor(int i); // temporary solution, only to test layers
 	};
 
+	class Rectangle : public Shape {
+
+	public: 
+		Rectangle(int layer = 0);
+	};
 }
 
 #endif
